feat: rename `speckle` to `specklepy`
