+++ specklepy/objects/units.py
@@ -1,4 +1,4 @@
-from speckle.logging.exceptions import SpeckleException
+from specklepy.logging.exceptions import SpeckleException
 
 UNITS = ["mm", "cm", "m", "in", "ft", "yd", "mi"]
